functions-etl/requirements.txt
azure-functions
requests
lxml
psycopg2-binary
azure-storage-blob